Increase extensibility of shadow system
